+++ src/ScriptUtils/cli.ts
@@ -7,11 +7,16 @@
  */
 
 import { Command } from 'commander';
+import { tsImport } from 'tsx/esm/api';
 
 import type { MaybePromise } from '../Async.ts';
 
 import { invokeAsyncSafely } from '../Async.ts';
-import { getDirname } from '../Path.ts';
+import {
+  getDirname,
+  join,
+  relative
+} from '../Path.ts';
 import {
   buildClean,
   buildCompile,
@@ -29,11 +34,17 @@ import {
 } from './esbuild/ObsidianPluginBuilder.ts';
 import { lint } from './ESLint/ESLint.ts';
 import { format } from './format.ts';
-import { process } from './NodeModules.ts';
+import {
+  existsSync,
+  process
+} from './NodeModules.ts';
 import { readPackageJson } from './Npm.ts';
 import { publish } from './NpmPublish.ts';
+import { ObsidianDevUtilsRepoPaths } from './ObsidianDevUtilsRepoPaths.ts';
+import { resolvePathFromRootSafe } from './Root.ts';
 import { spellcheck } from './spellcheck.ts';
 import { updateVersion } from './version.ts';
+
 /**
  * The number of leading arguments to skip when parsing command-line arguments.
  * The first two elements typically represent the Node.js executable and the script path:
@@ -59,6 +70,11 @@ enum CommandNames {
   Publish = 'publish',
   Spellcheck = 'spellcheck',
   Version = 'version'
+}
+
+interface OverrideModule<Args extends unknown[]> {
+  // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+  invoke(...args: Args): Promise<CliTaskResult | void>;
 }
 
 /**
@@ -119,5 +135,20 @@ function addCommand<Args extends unknown[]>(
 ): Command {
   return program.command(name)
     .description(description)
-    .action((...args: Args) => wrapCliTask(() => taskFn(...args)));
+    .action((...args: Args) =>
+      wrapCliTask(async () => {
+        const scriptPath = resolvePathFromRootSafe(join(ObsidianDevUtilsRepoPaths.Scripts, `${name.replace(':', '-')}.ts`));
+        if (existsSync(scriptPath)) {
+          const dir = getDirname(import.meta.url);
+          const relativeScriptPath = relative(dir, scriptPath);
+          const module = await tsImport(relativeScriptPath, { parentURL: import.meta.url }) as Partial<OverrideModule<Args>>;
+          if (typeof module.invoke !== 'function') {
+            throw new Error(`${relativeScriptPath} does not export an invoke function`);
+          }
+          return module.invoke(...args);
+        }
+
+        return await taskFn(...args);
+      })
+    );
 }